Add two new analyzers, remove the old
