+++ 2333/2/versus.rb
@@ -1,15 +1,16 @@
 require_relative 'libs/options_parser.rb'
 require_relative 'libs/handler.rb'
-require_relative 'libs/analyzer.rb'
+require_relative 'libs/top_bad_words_analyzer.rb'
+require_relative 'libs/top_words_analyzer.rb'
 
 parameters = ParametersParser.new.options
 rappers = Handler.create_rappers_array
 
 if parameters[:top_bad_words]
-  versus_analyzer = Analyzer.new(rappers, parameters[:top_bad_words])
-  versus_analyzer.top_bad_words
+  analyzer = TopBadWordsAnalyzer.new(rappers, parameters[:top_bad_words])
+  analyzer.top_bad_words
 elsif parameters[:name]
   put_count = parameters[:top_words] || 30
-  versus_analyzer = Analyzer.new(rappers, put_count, parameters[:name])
-  versus_analyzer.top_words
+  analyzer = TopWordsAnalyzer.new(rappers, put_count, parameters[:name])
+  analyzer.top_words
 end
